Make role details page index page
